+++ core/specfem/kokkos_kernels.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "kokkos_kernels/domain_kernels.hpp"
+#include "kokkos_kernels/frechet_kernels.hpp"
