+++ core/specfem/kokkos_kernels/impl/compute_material_derivatives.tpp
