+++ core/specfem/kokkos_kernels/impl/compute_seismogram.tpp
