+++ core/specfem/kokkos_kernels/impl/compute_stiffness_interaction.tpp
