+++ core/specfem/runtime_configuration/solver.tpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "kokkos_kernels/domain_kernels.hpp"
+#include "specfem/kokkos_kernels.hpp"
 #include "specfem/solver.hpp"
 #include "specfem/solver.hpp"
 #include "specfem/timescheme/newmark.hpp"
