+++ docs/sections/api/specfem/kokkos_kernels/compute_mass_matrix.rst
