+++ docs/sections/api/specfem/kokkos_kernels/compute_material_derivatives.rst
