+++ docs/sections/api/specfem/kokkos_kernels/invert_mass_matrix.rst
